requirements.txt
streamlit
openai
azure-storage-blob
PyPDF2
langchain
ragas
python-dotenv
faiss-cpu